pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "friesa-k-crai2026"
version = "1.0.0"
description = "Artifact for: From Agent Failure Paths to Quantified Residual Risk — CRAI @ DSN 2026"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.6,<3",
  "pyyaml>=6.0,<7",
  "numpy>=1.26,<3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
